base refactoring complete
need to make consistent styling
