+++ src/Core.jsx
@@ -17,7 +17,7 @@ import {
 	DatabaseView,
 	DatabasesView,
 	SearchResultsView,
-} from './db-views'
+} from './views/db-views'
 import './global.css'
 
 const PRIMARY_COLOR = '#6247aa'
