+++ src/components/db-components/Systems.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Box, Typography, Stack, Chip } from '@mui/material'
 import { useParams } from 'react-router-dom'
 import { initIPFS, initOrbitDB, getAllDatabases } from '../../services/database'
-import { actions, useStateValue } from '../../state'
+import { actions, useStateValue } from '../../state/DatabaseState'
 
 export function Systems() {
 	const [appState, dispatch] = useStateValue()
