+++ src/components/db-components/controls/CounterStoreControls.jsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react'
 import { Box, Button, TextField, InputLabel } from '@mui/material'
 import AddIcon from '@mui/icons-material/Add'
 
-import { useStateValue, actions } from '../../../state'
+import { useStateValue, actions } from '../../../state/DatabaseState'
 
 export default function CounterStoreControls() {
 	const [appState, dispatch] = useStateValue()
