+++ src/components/db-components/controls/FeedStoreControls.jsx
@@ -10,7 +10,7 @@ import {
 } from '@mui/material'
 import AddIcon from '@mui/icons-material/Add'
 
-import { useStateValue, actions } from '../../../state'
+import { useStateValue, actions } from '../../../state/DatabaseState'
 
 export default function FeedStoreControls() {
 	const [appState, dispatch] = useStateValue()
