+++ src/components/db-components/controls/LogStoreControls.jsx
@@ -8,7 +8,7 @@ import {
 	MenuItem,
 } from '@mui/material'
 import AddIcon from '@mui/icons-material/Add'
-import { useStateValue, actions } from '../../../state'
+import { useStateValue, actions } from '../../../state/DatabaseState'
 
 export default function LogStoreControls() {
 	const [appState, dispatch] = useStateValue()
