+++ src/views/Dash.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Outlet, useParams, useNavigate } from 'react-router-dom'
 import { Box } from '@mui/material'
 import { Contract } from 'near-api-js'
-import { ProjectDetailsContext } from '../ProjectDetailsContext'
+import { ProjectDetailsContext } from '../state/ProjectDetailsContext'
 import Navigation from '../components/Navigation'
 
 export default function Dash() {
