+++ src/views/ProjectHome.jsx
@@ -4,7 +4,7 @@ import GitHubIcon from '@mui/icons-material/GitHub'
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever'
 import { useNavigate } from 'react-router-dom'
 import Swal from 'sweetalert2'
-import { ProjectDetailsContext } from '../ProjectDetailsContext'
+import { ProjectDetailsContext } from '../state/ProjectDetailsContext'
 import ConfigFile from '../components/ConfigFile'
 import CodeImage from '../assets/code.svg'
 import { deleteNEARProject } from '../services/NEAR'
